Put duplicated functors in an extra file
